+++ react/hooks/useOrderFormCustom.ts
@@ -1,7 +1,12 @@
+import { useEffect, useState } from 'react'
 import { useQuery } from 'react-apollo'
 import type { OrderForm as OrderFormType } from 'vtex.checkout-graphql'
 import { OrderForm } from 'vtex.order-manager'
-import type { OrderForm as OrderFormSeller, Query } from 'vtex.store-graphql'
+import type {
+  Query,
+  Address,
+  OrderForm as OrderFormSeller,
+} from 'vtex.store-graphql'
 
 import GET_ORDER_FORM_SELLERS from '../graphql/getOrderFormSellers.graphql'
 
@@ -9,13 +14,19 @@ const { useOrderForm } = OrderForm
 
 type OrderFormQuery = Pick<Query, 'orderForm'>
 
+type PaymentAddress = {
+  paymentAddress?: Address | null
+}
+
 export type UseOrderFormReturn = {
   loading: boolean
-  orderForm: OrderFormType & Pick<OrderFormSeller, 'sellers'>
-  setOrderForm: (orderForm: OrderFormType) => void
+  orderForm: OrderFormType & Pick<OrderFormSeller, 'sellers'> & PaymentAddress
+  setOrderForm: (orderForm: OrderFormType & PaymentAddress) => void
 }
 
 export function useOrderFormCustom() {
+  const [paymentAddress, setPaymentAddress] = useState<Address | null>(null)
+
   const { data, loading: sellersLoading } = useQuery<OrderFormQuery>(
     GET_ORDER_FORM_SELLERS,
     {
@@ -31,9 +42,15 @@ export function useOrderFormCustom() {
     setOrderForm,
   } = useOrderForm() as UseOrderFormReturn
 
+  useEffect(() => {
+    if (orderForm?.shipping?.selectedAddress) {
+      setPaymentAddress(orderForm.shipping.selectedAddress)
+    }
+  }, [orderForm])
+
   return {
     loading: loading || sellersLoading,
-    orderForm: { ...orderForm, sellers },
+    orderForm: { ...orderForm, sellers, paymentAddress },
     setOrderForm,
   }
 }