+++ react/utils/messages.ts
@@ -1,6 +1,21 @@
 import { defineMessages } from 'react-intl'
 
 export const messages = defineMessages({
+  BRA: { id: 'store/checkout.b2b.country.BRA' },
+  USA: { id: 'store/checkout.b2b.country.USA' },
+  FRA: { id: 'store/checkout.b2b.country.FRA' },
+  cancel: { id: 'store/checkout.b2b.cancel' },
+  confirm: { id: 'store/checkout.b2b.confirm' },
+  editBillingAddress: { id: 'store/checkout.b2b.modal.editBillingAddress' },
+  billingAddress: { id: 'store/checkout.b2b.totalizer.billingAddress' },
+  street: { id: 'store/checkout.b2b.address.street' },
+  number: { id: 'store/checkout.b2b.address.number' },
+  complement: { id: 'store/checkout.b2b.address.complement' },
+  neighborhood: { id: 'store/checkout.b2b.address.neighborhood' },
+  city: { id: 'store/checkout.b2b.address.city' },
+  state: { id: 'store/checkout.b2b.address.state' },
+  postalCode: { id: 'store/checkout.b2b.address.postalCode' },
+  country: { id: 'store/checkout.b2b.address.country' },
   shippingAddress: { id: 'store/checkout.b2b.totalizer.shippingAddress' },
   shippingOption: { id: 'store/checkout.b2b.totalizer.shippingOption' },
   shippingOptionEmpty: {
@@ -10,6 +25,7 @@ export const messages = defineMessages({
     id: 'store/checkout.b2b.totalizer.shippingOption.packages',
   },
   emptyAddress: { id: 'store/checkout.b2b.totalizer.emptyAddress' },
+  editAddress: { id: 'store/checkout.b2b.totalizer.editAddress' },
   total: { id: 'store/checkout.b2b.totalizer.total' },
   paymentMethods: { id: 'store/checkout.b2b.totalizer.paymentMethods' },
   selectPaymentMethods: {